Add azimuth equal-area projection
The distance of an object from the centre of the screen is proportional to that angle's difference on a sphere from the current angle (through the sphere instead of its surface, no longer linearly proportional to the angle you must rotate to face it), which counteracts the objects near the edges being wider (though they're rendered as circles regardless, so it only affects their packing (denser radially, proportionally to the amount they're sparser perpendicularly)).

Now the default but you can switch between it and the previous azimuth equidistant one with the p key (the scale is almost exactly the same, but objects near the centre appear to be expanded outwards and near the edge compressed), I think it's better but will continue experimenting.

The projection doesn't expand distances itself, though, it's scaled such that the distance from the origin at 90º field of view (from the origin to the screen corners) is the same. The conversion factor for distance is sqrt(sin(m)^2+(1-cos(m))^2)/m, it works in radians, degree versions are in the initialisation for field of view conversion.