Add hex lattice mode for low-resolution raytracing
I realised that since the projection of a ray takes much more computing power than the setting of the pixel's initial angle, arranging them better on the screen could increase the effective resolution per rays traced.
In a lattice of square unit pixels, the average distance (from a simulation of 50 million) of a random point from the centre of the pixel it's closest to is ~0.3825. In one of hexagonal pixels, it's ~0.3510, a decrease of near 10%, or an increase of near 9% resolution. (See "Simulate Lattice Resolution" function, cyan pixels are those it measures, red ones are either out of bounds (it only measures a triangular segment) or closer to another one.)
Therefore, for the 96*72 live raytracing mode, if it were to instead iterate through 2/sqrt(3)*h*4/3*h+h/2 pixels (first term adjusted height, second is width, third is extra edge ones in offset rows) instead of h*4/3*h, for 9% more resolution per pixel.
https://www.wolframalpha.com/input/?i=y%3D%282%2Fsqrt%283%29*h*4%2F3*h%2Bh%2F2%29%2F0.3510%2C+y%3D%28i*4%2F3*i%29%2F0.3825%2C+y%3D%2872*4%2F3*72%29%2F0.3825&assumption=%22i%22+-%3E+%22Variable%22
The hex lattice needs a height (before corrugation) of 64.0231 to reach the same resolution as the square one needs 72 for, meaning 6342.77 rays (according to the maths) and 6,401 (according to the implementation with the rounding) traced instead of 6912, meaning a performance increase of almost 8% for no resolution difference.
Press p on startup for hexagonal raytracing mode, or p and k simultaneously for square, or t at any time to stop and do high-quality raytracing.