Add dot product collisions to physics collisions as well as raytracing ones
Now seems much more accurate to real physics (because balls' motion in the lateral plane doesn't invert, only parallel motion (which swaps in the specific case of equal-mass billiards balls (which is also why, in 8-ball (as seen in my previous overhead 2D 8-ball program) the resultant trajectories after a collision with a stationary ball are perpendicular))).
Also fix minor pixel-offset issues in raytracing (where the rays' screen positions were minorly offset), and also increase height of hexagonal lattice mode by one row (because it previously left constant-colour gaps when higher-resolution screens left behind it weren't cleared, and I don't want to add automatic clearing because this way the previous frame is rendered in the bottom half of the screen when the current frame is only partially rendered).