Add Minkowski space (in progress) and stereographic projection (projection mode 4)
The Minkowski mode is an extension of the finite light speed mode in raytracing, where the time rays take to reach objects is finite, as well as their relative velocities, so closer spheres (and closer pixels in spheres) will appear to be closer to the present, and further ones will be further in the past (or the extrapolated past, where it doesn't compute collisions so spheres phase through each other (with interesting, non-circular intersection boundaries)).
In stereographic mode, currently the non-raytracing mode doesn't change size correctly, things at the centre that are closest to you can obscure others behind them, but it works properly in raytracing mode. What is 'in front' of you (according to the WASD controls) is on the outside of the screen, because the inside is small and the real-time mode makes it pixellated beyond the point of usefulness.
In stereographic map projections, infinitesimal shapes aren't scaled relatively on either side more than the other, so their shape is retained. In mapping planets where preserving orientation (due to poles and spin pivot axes) and area are important, the Mercator reduces distortion (https://www.physics.drexel.edu/~goldberg/projections/goldberg_gott.pdf, table on page 31, better in terms of area, flexion, skewness and distance), but when your planet (or unit sphere to which 3D is projected) has many circles on it (which spheres are projected to), if you want to retain the shape of the boundaries (like if the regular mode is only efficiently able to render circles), the stereographic one is better, it will allow a more confusing non-raytracing mode but which doesn't have the current issues with the Lambert azimuthal equi-area one where spheres appear to phase through each other (because the circles seen will be their boundaries).
Because the centre of the sphere is closer to the origin than the centre of the projection (as seen by the skewedness of the brightness rings (proportional to collision magnitude)), the projection's centre and radius will be found instead by averaging the closest and furthest points.
Also, all raytracing projections converted from screen coordinates to absolute angle only after performing angular magnitude trigonometry, now they do it before (simplifying it somewhat), not sure whether this has any effect but if it does, all but the azimuthal equidistant were wrong until now.