Add stereographic non-raytracing mode
More difficult than the raytracing, because you need to average the positions in the projected circles closest and furthest from the screen origin to get the projected centre (further away than the actual one) and subtract one from the other for the projected radius, instead of only applying the equation from Wikipedia given a ray's screen position.
In full resolution, this mode will make you dizzy and seasick, size no longer corresponds with distance and no Euclidean intuition works. Remember, forwards (the w key) is the outside of the screen (because it would otherwise be too small), backwards is the inside, don't rotate too much.
(It already existed before this commit, but projected spheres at the locations and radii of their non-projected centres, making it unusable, because spheres near the centre of the screen and close to the camera obscured most others, but now it's entirely geometrically accurate, all apparent flickering from distance order changing between two intersecting each other in stereographic is from physics, not the projection.)
New keyboard controls also, 1 is azimuthal equidistant, 2 equi-area, 3 stereographic, 4 will be gnomonic, 0 is for original wacky one I made accidentally, with the squares attached at vertices.