Add refractive glass balls
Implementation is based on https://raytracing.github.io/books/RayTracingInOneWeekend.html#dielectrics/snell'slaw (no dielectrics yet), it doesn't look right (objects' images are elongated instead of truncated towards its edges, but this effect seems to reduce when viewed from further away, so it may only be an artefact of focal distance). It still abides by the rule (in the program's wacky physics) of colour change imparted to the ray being proportional to the magnitude of the change in velocity, so the centre is transparent and the edges opaque.

There is now a large glass ball in the demonstration scene in front of the others, which now have domain colouring on their initial positions instead of being random (where x is red, y is green and z is blue, they can't be too extreme in any colour channel without being at 127 in the others because they're arranged in a sphere) to show more clearly that the refracted image is upside-down without having to look at individual balls' locations.

Also fix orthogonal mode (balls were previously rendered at variable size based on whichever was closest to the perspective camera, meaning some would occupy the entire screen and flash).

In the next commits, I will add diffusive balls (there won't be antialiasing (at least not in the real-time mode (because if it could afford to antialias it could afford to increase resolution)), so each pixel will flicker to reflecting different objects); retroreflector balls (I think they'd be physically possible in real life if you had a cubic lattice of mirrors triagonal to the surface), which would appear to be a constant brightness throughout (just like real ones) but which would display each ray's reflections throughout its course in reverse inside their reflections in it; eventually a solar system mode with an orbital velocity function like in my 8-ball program's orbital mode. The solar system would probably be our one but downscaled (enough to make it interesting to knock planets around and travel long distances reasonably easily), with proper lighting (where planets only subtract light from rays that doesn't match their own hue, and in each ray's final unobstructed collision the origin light comes from the sun).

Also, this commit contains secret collision mode 3 (will not be documented, is only for ball-ray collisions (not ball-ball ones), and is to be used only by dot product enthusiasts) where it uses the dot product's components instead of multiplying its absolute velocity by the components of the position offset at the moment of the collision (causing it to behave differently at different orientations and have a cool banding effect around the orthogonal).