Hotfix: Allow proper refraction for intersecting materials, fix Minkowski mode (properly this time)
Previously there was only a variable that would alternate each refraction to represent whether the ray is inside a ball (so that it can use the reciprocal of the refractive index for when it exits), now there's a list of balls the ray is inside instead to support intersecting ones, you would think this would be shallow and pedantic but it will support hollow ones (two concentric spheres (one of sub-1 refractive index (the raytracing in one weekend book says to use negative radius but then they would have to be added to the list by default)), where balls inside the smaller one are bounced in sub-frames to remain inside it (the special case of the shell theorem would mean no gravity would be imparted on them) and they would roll in interesting ways).
It turns out all of the problems in the Minkowski mode were from the balls' velocities (for their projected positions) using the o counter variable instead of candidate o (and I had made the c [x,y,z,i,j,k] variables for conciseness (because their ball list elements would only be referenced once when they're made the candidate), but hadn't done what would have also been obvious if I were trying to compact the Minkowski programming instead of fix it). However, now it works (will make interface for configuring projections and light speed and such soon).