Add preliminary version of Python rewrite
I had been making an early version, the "not very working Python version" in my 8-ball repository (which, despite the repository, was based on the Scratch program here), but couldn't get it to work and gave up. However, after having made my tablebase program and become better at Python, I decided to have another go. The collision functions are ported from the Scratch one here (and cleaned up greatly), but the rest of the physics engine (with the perspective 3D and quaternion rotation) is from the tablebase program's state transition diagram mode (an improved version of some parts of the one in the 8-ball repository, which itself was loosely based on this).
I intend to eventually make some 2D things (perhaps even a game) out of this, so the computations all work with arbitrarily many dimensions, so if you're 4-dimensional and have a Pygame port for your 4D displays' 3D surfaces, you only have to change dims to 4 and change the rendering functions to use this for watching n-body problems with hyperspheres, if you would like.