Make arm work with conservation of moments, add gravity domain colouring
After firing an arm, contract it by holding the mouse down, your lateral velocity is multiplied by the reciprocal of the ratio of the lengths in the previous and current frame, so you can do maneuvers to increase your speed. You can also now jump onto the underside of the smaller circles and move around to the top reasonably reliably now, as opposed to the previous commit where you wobbled uncontrollably, because your lateral acceleration is no longer affected by downwards acceleration to hang onto the circle (which was originally handled only by gravity and air resistance like the other modes).
The raytracing mode (toggled with the t key) in both 2D modes now renders gravity with domain colouring at each "pixel" (comprised of many screen pixels due to being otherwise slow).
This version has a great many problems, your velocity should be corrected from diverging from that of the circle you've landed on at a tangent, but it sometimes isn't quite, leading to jumps not registering when moving quickly (noticeable on the large one) and to falling off the smaller ones when on one side, and your position being changed by a contracting arm also allows you to phase through circles' edges and land within them, but it will be fun, I hope.