Make collision time determination work with differences in time as well as space and velocity
When a collision occurs (being the first one to do so in a frame), previously the positions of every node were updated to the current time in the subframe. However, their positions only need to be proceeded either at the end or when they're the ones colliding, and this commit adds equations (with the derivation) for the more complicated quadratic equation when the two colliding objects are stored at different times. I hope the reduced list reallocations copmensate for the increased complexity of the calculations.
Much seems cacheable, things like the absolute velocity squared of each node are recomputed each collision.
Also, trivially optimise by adding pairs that are never going to collide with their current trajectories to pairsAlreadyCollided.
Another optimisation can be made, by caching all pairs that are going to collide at each subframe and later removing those that include nodes that have collided, and reusing the cache instead of recomputing each subframe.
This optimisation could also be ported to the Scratch version if you wanted to.