make moving in direction of orientation work
It turns out that spatial position and orientation in spherical 3-space (equivalent to 4D hypersphere orientation) need to be intertwined in versors together (one enacted by multiplication on either side), which can be derived from the equation for rotating components in a 2D slice without affecting components in the perpendicular plane (instead of perpendicular line (ie. cross product), because it's entirely rotating the hypersphere directly instead of 3D space afterwards).
It is no longer a cool one-liner like before, but allows some parts to be more elegant (ie. handling everything before the logarithm), I struggled for a while before having the idea to multiply by the rotation versor on one side and its conjugate on the other (which seems to work (the best kind of working))
Pairs of versors are also the minimal, information-wise
The only problem is that velocity is rotated with the camera's orientation, however, but for the time being consider this a feature (to reduce motion sickness or something)
(I'm committing this version, similarly to the previous one, as the first one that I got working with a change, for documentation purposes)

You also may notice the other mode in rotationParameters (which is for enacting it directly upon a value x, in a manner less composable than returning two new quaternions that the existing pair can be multiplied by, but in only 64 numerical multiplications (instead of the 40 to generate the quaternion pair and 32 more* to multiply through), using the fact that many terms of the expansion measure absolute values of specific quaternion parameters, which are 1 when they're versors**), which is there because I have begun using SymPy*** to help me (which doesn't have any equation length constraints like WolframAlpha (and isn't paid like Mathematica)), I like it a lot and will be able to do things with complicated derivations (like the original 3D engine's mode with storing time offsets from the "present" in a sub-frame for each cell also) more easily (ie. no longer splitting them apart to attempt to simplify piecewise with WolframAlpha) when they arise
*There is https://en.wikipedia.org/wiki/Multiplication_algorithm#Complex_number_multiplication that accomplishes complex multiplication in three numerical multiplications (Strassen-style), I wonder whether there's anything like that for quaternions
**no numerical stability, albeit (living life on the edge)
***(also used in my tablebase program for multiset permutations but I didn't know about its general CAS functionality having a WolframAlpha-like plaintext interface)
Anyway, I have big plans for it still