All manner of wondrous things
Sorry for my inactivity as of late, I have been quite busy with real-life preoccupations once more, but I have a few additions of various degrees of whimsy and interestingness (though nothing that can be seen without changing parts of the program).

I had previously made a more complete algebra of quaternions than had been necessary, however now I have made one of permutations also, the idea is that they are meant to contain the terms on which they act, and leave all others as-is, and compose by multiplication (which explains the factoradic indexing methods (as a treat), and the brief foray into implementing some OEIS sequences seemingly entirely unrelated to 4D engines for the purposes of their fast exponentiation, involving bounds for their periods (that turned out not to be necessary because converting to cycle format is fast enough as-is (as you will find out if (for whatever reason) you want to make a table of all permutations up to 7! to the power of their index))).

Perhaps I will remove all of the broken and commented-out permutation __int__ methods (from before I had my revelation that (along with a significant amount of calibration by experimentation) led to the working one), but I like some, especially those that generate palindromes delimited by zeros (very mysterious).

There is also no longer a 120*120 multiplication table stored inline in the program (though it looks quite nice (with many interesting patterns in the numbers of digits of adjacent numbers, due to the decision for axial permutations to take precedence over signs in the ordering), if you want to look at it as it was then), because there are now some functions strewn about for finding minimal sets of rootless ('prime') actions, of which recursively taking the set of compositions of their Cartesian products with themselves eventually yields the entire initiial set (which was the initial reason for the decision to implement the permutation class), I made a sorted version of product (sortduct) before learning it was already implemented as combinations_with_replacement.

I also added a set of quaternions for orientations in the icosahedral group (after realising that it could be implemented by rotations by 2*pi/5 in the roll axis as well as across edges between orientations pointing towards vertices (imparting there also a roll of pi), so they generate to a finite set instead of growing uncontrollably). Several ways to express them are provided (in terms of 5, phi (that some people like, and reduces the number of sqrts necessary by 1 in all cases (though so too would assigning sqrt(5) as a constant)) and roots of quadratics), though the surd class wasn't made with nested roots in mind (and their canonicalisation seems quite hard) so I decided to instead compute them numerically (and use Robert Munafo's RIES instead of calculating them all manually (which went well, only minor adjustments were necessary in some cases)). Their generation from the program is quite elegant, I think, and I will be able to make the rotation group of the 600-cell next (being that this is the vertex figure).

Perhaps it would be a good idea to make a quaternion pair class, also (being that they come up in multiple places now, and the x!=None case in rotationParameters provides a way (found with SymPy and experimentation) for generating 4D rotation matrices out of the same parameters as them).

I also made some camera-rotation functions at the end (mainly for testing), and made the edges do spherical-linear-interpolations (slerps) with precomputed roots, so you can truly feel as though you're in moving around edges inscribed in the surface of a hypersphere (though it doesn't matter much, because lines still render behind you).

I also got antipodal points rendering more dimly than their nearby counterparts then broke it again, many of the extraneous additions towards the end are from me trying to fix it again, sorry (I ought to commit more and program when more awake).

Anyway, I hope you, my dear perusers (of the present and future) will enjoy, I unfortunately must stop distracting myself with all this productive waste of time for the time being once more, though I will return to finish it eventually.